Make changes based on code review
